+++ tests/testthat/test-class-command-args.R
@@ -107,7 +107,6 @@ test_that("$add_argument(action = 'flag') [#17]", {
   expect_identical(obj, exp)
 
   # nolint start: line_length_linter.
-  expect_warning(command_args()$add_argument("f", action = "flag", default = TRUE))
   expect_warning(command_args()$add_argument("f", action = "flag", default = "1"))
   expect_warning(command_args()$add_argument("f", action = "flag", default = "1"))
   # nolint end: line_length_linter.