+++ NEWS.md
@@ -1,6 +1,7 @@
 # scribe (development version)
 
 * positional arguments now can have default values [#52](https://github.com/jmbarbone/scribe/issues/52)
+* `scribeArgs` with `action = 'flag'` now accept `default = TRUE` [#55](https://github.com/jmbarbone/scribe/issues/55)
 
 # scribe 0.1.0